Folder restructure works
