+++ src/collectors/__init__.py
@@ -1,0 +1,2 @@
+from .factory import Collector_Factory, Collector_Types
+from .collectors import Collector
